requirements.txt
# Fireflies Transcripts RAG Pipeline Requirements

# Core dependencies
python-dotenv==1.1.1
requests==2.32.4
supabase==2.18.1
openai==1.99.9
tiktoken==0.11.0

# Required for optimized pipeline
numpy==2.0.2  # For vector operations

# Optional dependencies for webhook server
fastapi==0.100.0
uvicorn==0.23.0